# Content import was too slow. Now it is less slow.