… kolibri/content/fixtures/content_test.json
… kolibri/content/test/test_channel_import.py
… kolibri/content/test/test_content_app.py
… kolibri/content/test/test_import_export.py
… kolibri/core/analytics/management/commands/ping.py
… kolibri/plugins/device_management/assets/src/state/actions/contentWizardActions.js
… kolibri/plugins/device_management/assets/src/state/actions/manageContentActions.js
… kolibri/utils/options.py