<!--
 1. Following guidance below, replace …'s with your own words
 2. After saving the PR, tick of completed checklist items
 3. Skip checklist items that are not applicable or not necessary
 4. Delete instruction/comment blocks
-->

### Summary
<!--
 * description of the change
 * manual verification steps performed
 * screenshots if the PR affects the UI
-->

Speeds up the following operations:
- Loading the content page (the list of existing channels): 6x SPEEDUP
- Importing channel metadata ("preparing channel listing"): 10x SPEEDUP
- Importing the content itself (downloading content files): 23x SPEEDUP

### Reviewer guidance
<!--
 * how can a reviewer test these changes?
 * are there any risky areas that deserve extra testing
-->

Test the process of importing new channels and content, and compare speeds between doing so with the current 0.10 branch and this code. Will be especially noticeable with large channels like Khan Academy (English).

Note: need to resolve some issues raised by tests tomorrow, with @rtibbles.

### References
<!--
 * references to related issues and PRs
 * links to mockups or specs for new features
 * links to the diffs for any dependency updates, e.g. in iceqube or the perseus plugin
-->

```
Channel listing page for import (with several channels loaded):

    Before:
        Calculating all 4 values: 12s <- what we currently do on channel list page
        Calculating size on disk: 3.5s

    After:
        Calculating all 4 values: 7s
        Calculating size on disk: 2s <- what we do now on the channel list page

=============================

Old Metadata Import  <- what we did before
-------------------
Khan Academy: 2m44s
Engage NY: 18.5s

New Metadata Import  <- what we do now
-------------------
Khan Academy: 20s (8x speedup)
Engage NY: 1.5s (12x speedup)

=============================

Downloading Channel over HTTP
-----------------------------
3m45s, 4m3s, 2m54s, 4m42s, 5m41s
Average: 253s (slowdown due to 301 redirects)  <- what we did before
(When done through UI, 3 samples: 199s)

Downloading Channel over HTTPS
------------------------------
2m12s, 2m5s, 1m0s, 1m46s, 1m45s
Average: 106s (2.4x speedup)
(When done through UI: 117s)

Downloading Channel over HTTPS with Shared Session
--------------------------------------------------
12s, 10s, 10s, 11s, 11s
Average: 11s (23x speedup)  <- what we do now
(When done through UI: 12s)
```


### Contributor Checklist

- [ ] Contributor has fully tested the PR manually
- [ ] PR has the correct target branch and milestone
- [ ] PR has 'needs review' or 'work-in-progress' label
- [ ] If PR is ready for review, a reviewer has been added. (Don't use 'Assignees')
- [ ] If there are any front-end changes, before/after screenshots are included
- [ ] If this is an important user-facing change, PR or related issue has a 'changelog' label

### Reviewer Checklist

- [ ] Automated test coverage is satisfactory
- [ ] Reviewer has fully tested the PR manually
- [ ] PR has been tested for [accessibility regressions](http://kolibri-dev.readthedocs.io/en/develop/manual_testing.html#accessibility-a11y-testing)
- [ ] External dependencies files were updated (`yarn` and `pip`)
- [ ] Documentation is updated
- [ ] Link to diff of internal dependency change is included
- [ ] CHANGELOG.rst is updated for high-level changes
- [ ] Contributor is in AUTHORS.rst