@lyw07 2018-06-25
sorry what is a good way to test the speed? Thank you!