@radinamatic 2018-06-25
Well, this thing certainly flies like a Kolibri, and no stopwatch necessary... 🚀 🎉 💥 
15 minutes download time on 0.10beta5 reduced to 3 minutes in this build!!!

Tested with PEX on Ubuntu 16.04, as Windows installer gives the same error as in #3852, installed channels in various languages, checked that resources display correctly in channels, recommended, lessons & exams for both learners, coaches and admins. Nothing seems to fail, everything is just so beautiful and fast! 😍 

![2018-06-26_1-18-26](https://user-images.githubusercontent.com/1457929/41881171-646803ea-78e2-11e8-8109-b564e3296c2e.png)

![2018-06-26_1-28-16](https://user-images.githubusercontent.com/1457929/41881181-6b58ad4e-78e2-11e8-92f1-de9214a61556.png)

![2018-06-26_1-30-01](https://user-images.githubusercontent.com/1457929/41881191-73391cf6-78e2-11e8-8be8-0e857596888b.png)