@indirectlylit 2018-06-27
fixes #3866